String Hashing

+++ Code Templates/binary_exp.cpp
@@ -6,8 +6,8 @@ class BinaryExponentiation {
 
     public:
     //If a, b, mod <= 1e9
-ll myPow(ll a, ll b) {
-    ll ans = 1;
+int myPow(int a, int b) {
+    int ans = 1;
     while(b) {
         if(b % 2) {
             ans = (ans * a) % mod;
